fix: 全面修复所有 8 个编译错误
- 删除 TicketGrabbingStrategy.kt 中重复的 APIGrabbingStrategy 类
- 修复 PreciseTimeManager.getTimeSyncStatus() 调用方式
- 添加 DeviceFingerprint 数据类缺失的 language 字段
- 添加 generateRandomLanguage() 方法
- 修复 HumanBehaviorSimulator.kt 中 Double 到 Float 的类型转换
- 添加缺失的 cos 和 ln 函数导入

+++ app/src/main/java/com/damaihelper/utils/HumanBehaviorSimulator.kt
@@ -8,6 +8,8 @@ import android.util.Log
 import android.view.accessibility.AccessibilityNodeInfo
 import kotlinx.coroutines.delay
 import kotlin.math.abs
+import kotlin.math.cos
+import kotlin.math.ln
 import kotlin.math.pow
 import kotlin.math.sqrt
 import kotlin.random.Random
@@ -197,8 +199,8 @@ class HumanBehaviorSimulator {
 
             for (i in 1 until pointCount) {
                 // 添加随机扰动（高斯分布）
-                val perturbationX = random.nextGaussian() * 20
-                val perturbationY = random.nextGaussian() * 20
+                val perturbationX = (random.nextGaussian() * 20).toFloat()
+                val perturbationY = (random.nextGaussian() * 20).toFloat()
                 
                 currentX += dx + perturbationX
                 currentY += dy + perturbationY
